refactor(domain): BridgeGame 패키지 이동, 관련된 접근제어자 변경
domain -> service
default -> public
